Extract components from App.tsx
