+++ src/components/SearchIconWrapper.tsx
@@ -1,0 +1,11 @@
+import {styled} from "@mui/material/styles";
+
+export const SearchIconWrapper = styled("div")(({theme}) => ({
+    padding: theme.spacing(0, 2),
+    height: "100%",
+    position: "absolute",
+    pointerEvents: "none",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+}));
